requirements.txt
screeninfo
openai
pynput
pillow
